+++ ChangeLog.md
@@ -98,6 +98,10 @@ was used.  Thus, this issue did not cause a segfault or other user-visible
 errant behavior (it was only detectable with ASan), and it did not likely pose
 a security risk.
 
+14. Hardened the TurboJPEG API against hypothetical applications that may
+erroneously call `tjCompress*()` or `tjTransform()` with a reused JPEG
+destination buffer pointer while specifying a destination buffer size of 0.
+
 
 2.1.5.1
 =======